First bugfree rework version
