+++ server/.env.template
@@ -28,15 +28,11 @@ json_schema="."
 json_text_content=False
 xml_xpath="//"
 
-max_document_limit=10
-vector_store=milvus
-vector_store_uri='data.db'
-vector_store_collection=ragmeup_documents
-vector_store_sparse_uri=bm25_db.pickle
-vector_store_initial_load=True
+postgres_uri="postgresql+psycopg://langchain:langchain@localhost:6024/langchain"
+postgres_collection=ragmeup_documents
+
+vector_store_load_file="load.signal"
 vector_store_k=10
-document_chunks_pickle=rag_chunks.pickle
-file_upload_using_llm=True
 rerank=True
 rerank_k=3
 rerank_model=flashrank
